Update
+ Added HeliSim config
+ Broke out ACE, ITC Air, Marker Lights and Reflectors into their own .hpp files
+ Deleted fza_ic_box.p3d
+ Deleted references to fza_ic_box in model.cfg
