+++ @AH-64D Apache Longbow/Addons/fza_ah64_controls/config/cfgVehicles/ace.hpp
@@ -1,0 +1,25 @@
+//Ace Hitpoints
+ace_repair_hitpointPositions[] = {
+    {"Hithull", 					{0,0,-1.5}},
+    {"HitEngine1", 					{-1,0.7,-1}},
+    {"HitEngine2", 					{1,0.7,-1}},
+    {"HitVRotor", 					{-0.6,-6.95,-0.1}},
+    {"HitHRotor", 					{0,2,0.7}},
+    {"HitTail", 					{0,-7,-2}},
+    {"HitVStabilizer1", 			{0,-7,-1}},
+    {"HitHStabilizerL1", 			{-2,1.8,-1.3}},
+    {"HitHStabilizerR1", 			{2,1.8,-1.3}},
+    {"Hitfuel", 					{0,1,-1.6}},
+    {"Hitfuel1", 					{0,5,-1.6}},
+    {"HitGun", 						{0,5,-2.3}},
+    {"HitTurret", 					{0,6.9,-1.5}},
+    {"PNVS", 						{0,6.8,-1}},
+    {"FCR", 						{0,2,1.6}},
+    {"TailBoom", 					{0,-3,-2}},
+    {"IrJammer", 					{-1,5,-1.6}},
+    {"RfJammer", 					{1,5,-1.6}},
+    {"Pylon1", 						{-2.385,2.4,-1.8}},
+    {"Pylon2", 						{-1.665,2.4,-1.8}},
+    {"Pylon3", 						{1.665,2.4,-1.8}},
+    {"Pylon4", 						{2.385,2.4,-1.8}}
+};
