+++ @AH-64D Apache Longbow/Addons/fza_ah64_controls/config/cfgVehicles/itcAir.hpp
@@ -1,0 +1,15 @@
+class itc_air 
+{
+    class rover 
+    {
+        capable = 1;
+        frequency_default = 5784;
+    };
+    targeting_user = "gunner";
+    hmd = 1;
+    tgp = 1;
+    wso = 0;
+    mfdApps[] = {"STAT","OPT"};
+    systems[] = {"UFC","MFD","MFD_R","ROVER","ACMI"};
+    mfdType = "classic";
+};
