+++ @AH-64D Apache Longbow/Addons/fza_ah64_us/model.cfg
@@ -1640,11 +1640,6 @@ class val##_y : val##_x {\
 			};
 		};
 	};
-	class fza_ic_box
-	{
-		sectionsInherit="";
-		sections[]={};
-	};
 	class fza_hydra_m261_wep
 	{
 		sectionsInherit="";
